successfully overloaded the add operator for integers
