+++ notes/TODO.md
@@ -59,3 +59,4 @@
 - POSSIBLE THINGS TO DO EVENTUALLY:
     - Accomplish Direct Threading by putting the 'next opcode fetch and jump' into a #define and
       putting it at the end of every opcode handler instead of a huge switch statement
+    - Get import working
